Fix github repo auth
This commit addresses the inherit platform differences involved
in the process of authenticating user credentials against a
repository when executing git commands that require
authentication, such as `git push`.

* .gitconfig: Remove setting a default credentials helper
script; this is the magic that allows us to deal with this
issue in a platform-specific mannor.

* .bashrc: On OSX, set GIT_ASKPASS to the appropriate helper script.
On Linux, do nothing and ...